+++ src/controller/ListUserController.ts
@@ -1,12 +1,25 @@
 import {ListUserRepository} from '../repositories/ListUserRepository';
 import {Request, Response, NextFunction} from 'express';
 import {Result} from '../utils/Result';
+import {Role} from '../enum/Role';
 
 export class ListUserController {
   private listUserRepository: ListUserRepository;
 
   constructor() {
     this.listUserRepository = new ListUserRepository();
+  }
+  async verifyIfUserIsAdmin(user_id: number, list_id: number) {
+    const user = await this.listUserRepository.findListUserByUserIdAndListId(user_id, list_id);
+    if (!user) {
+      return false;
+    }
+    return user.role === 'admin';
+  }
+
+  async verifyIfUserIsMember(user_id: number, list_id: number) {
+    const user = await this.listUserRepository.findListUserByUserIdAndListId(user_id, list_id);
+    return !!user;
   }
 
   async listUsersInList(req: Request, res: Response) {
@@ -39,6 +52,9 @@ export class ListUserController {
     if (!req.body.id || !req.body.role) {
       return Result.fail(400, 'Bad Request');
     }
+    if(!(await this.verifyIfUserIsAdmin(req.body.current_user_id, req.body.list_id))){
+      return Result.fail(401, 'Unauthorized');
+    }
     return this.listUserRepository.updateListUserRole(
       req.body.id,
       req.body.role
@@ -48,6 +64,9 @@ export class ListUserController {
   async addUserToList(req: Request, res: Response) {
     if (!req.body.user_id || !req.body.list_id) {
       return Result.fail(400, 'Bad Request');
+    }
+    if(await this.verifyIfUserIsMember(req.body.user_id, req.body.list_id)){
+      return Result.fail(409, 'User is already in list');
     }
     const listUser = {
       user_id: req.body.user_id,
@@ -60,6 +79,9 @@ export class ListUserController {
   async removeUserFromList(req: Request, res: Response) {
     if (!req.body.id) {
       return Result.fail(400, 'Bad Request');
+    }
+    if(!(await this.verifyIfUserIsAdmin(req.body.current_user_id, req.body.list_id))){
+      return Result.fail(401, 'Unauthorized');
     }
     return this.listUserRepository.deleteListUser(req.body.id);
   }
